+++ src/test/java/info/mackiewicz/bankapp/account/util/IbanGeneratorTest.java
@@ -1,15 +1,12 @@
 package info.mackiewicz.bankapp.account.util;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNotEquals;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-
+import info.mackiewicz.bankapp.shared.util.IbanValidationUtil;
+import info.mackiewicz.bankapp.testutils.TestIbanProvider;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
-import info.mackiewicz.bankapp.shared.util.IbanValidationUtil;
-import info.mackiewicz.bankapp.utils.TestIbanProvider;
+import static org.junit.jupiter.api.Assertions.*;
 
 class IbanGeneratorTest {
