+++ src/test/java/info/mackiewicz/bankapp/system/registration/service/DefaultBonusGrantingServiceTest.java
@@ -4,10 +4,10 @@ import info.mackiewicz.bankapp.account.model.Account;
 import info.mackiewicz.bankapp.account.service.AccountService;
 import info.mackiewicz.bankapp.shared.util.BankAccountProvider;
 import info.mackiewicz.bankapp.testutils.TestAccountBuilder;
+import info.mackiewicz.bankapp.testutils.TestIbanProvider;
 import info.mackiewicz.bankapp.testutils.TestUserBuilder;
 import info.mackiewicz.bankapp.transaction.model.Transaction;
 import info.mackiewicz.bankapp.transaction.service.TransactionService;
-import info.mackiewicz.bankapp.utils.TestIbanProvider;
 import org.iban4j.Iban;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
